requirements.txt
pytorch==1.12.0+cu113
cudatoolkit==11.3.1
tensorboard==2.6.0
tensorflow==2.6.0
pandas
sklearn
pyodbc
sqlalchemy
tqdm
numpy
scipy
seaborn
matplotlib
